project restructure for easier build. Added pytest. Cleaned up main.py
