+++ dit/__init__.py
@@ -1,0 +1,1 @@
+from .main import main
