+++ dit/core/Token.py
@@ -1,0 +1,6 @@
+# Token class. Determine how a token will look.
+class Token:
+    def __init__(self, type, value):
+        self.type = type
+        self.value = value
+        return
