+++ requirements.txt
@@ -1,14 +1,22 @@
+attrs==21.2.0
 click==8.0.3
+DataImportTool==0.0.1
 et-xmlfile==1.1.0
 greenlet==1.1.2
+iniconfig==1.1.1
 mysql-connector-python==8.0.26
 numpy==1.21.2
 openpyxl==3.0.9
+packaging==21.2
 pandas==1.3.3
+pluggy==1.0.0
 protobuf==3.18.1
+py==1.11.0
+pyparsing==2.4.7
+pytest==6.2.5
 python-dateutil==2.8.2
 python-dotenv==0.19.1
 pytz==2021.3
 six==1.16.0
 SQLAlchemy==1.4.25
-typer==0.4.0
+toml==0.10.2
